I did not do this previously because something wasn't ready. But now can't remember anymore why I didn't do it...

Obviously helps mostly for large files...