@rklaehn 2023-02-13
Speed comparison. Not very accurate since I just press Control-C as soon as I see the hash, but you get the idea...

main:
```
sendme on  main [?] is 📦 v0.1.0 via 🦀 v1.67.0 took 13s 
❯ time cargo run --release -- provide ../testdir_large_files/
    Finished release [optimized] target(s) in 0.13s
     Running `target/release/sendme provide ../testdir_large_files/`
Reading ../testdir_large_files/
Collection: bafkr4if7d6yhuz7xhmhjsd2z6v5sbgun5b6ciirimzwyazplzj3bwmlchq

- ../testdir_large_files/file1: 10 bytes
- ../testdir_large_files/test5: 1000000000 bytes

PeerID: TFNNvrnLXWflvoLultHM3QTG0V8oFs63Tw4HzbRW0Rc
Auth token: FoatKFDw3tT3d7gVMJfQ-y2qVS5J5QPMnjaHd5hL0oI
All-in-one ticket: IL8fsHpn9zsOmQ9Z9Xsgmo3ofCQiKGZtgGXrynYbMWI8IExTTb65y11n5b6C7pbRzN0ExtFfKBbOt08OB820VtEXAH8AAAHRIhaGrShQ8N7U93e4FTCX0PstqlUuSeUDzJ42h3eYS9KC
^C
Shutting down provider...
cargo run --release -- provide ../testdir_large_files/  6.53s user 0.74s system 90% cpu 8.011 total
```

this branch:
```
sendme on  multi-threaded-add [?] is 📦 v0.1.0 via 🦀 v1.67.0 took 3s 
❯ time cargo run --release -- provide ../testdir_large_files/
    Finished release [optimized] target(s) in 0.13s
     Running `target/release/sendme provide ../testdir_large_files/`
Reading ../testdir_large_files/
Collection: bafkr4if7d6yhuz7xhmhjsd2z6v5sbgun5b6ciirimzwyazplzj3bwmlchq

- ../testdir_large_files/file1: 10 bytes
- ../testdir_large_files/test5: 1000000000 bytes

PeerID: 0jSyRVTHtid2jvm-4B7rgKn3bt2Ol3fu4-TIkwTHp-8
Auth token: ABTVQeObBhXWZ0egjfcPqe3Xa6exLlitFWN0MhC1EHw
All-in-one ticket: IL8fsHpn9zsOmQ9Z9Xsgmo3ofCQiKGZtgGXrynYbMWI8INI0skVUx7Yndo75vuAe64Cp927djpd37uPkyJMEx6fvAH8AAAHRIgAU1UHjmwYV1mdHoI33D6nt12unsS5YrRVjdDIQtRB8
^C
Shutting down provider...
cargo run --release -- provide ../testdir_large_files/  6.87s user 1.12s system 350% cpu 2.280 total
```